refactor:  UserService interface 구현
